Interface

MOSInterface.Interface — Type

MOS interface struct.

MOSInterface.Interface — Method

Interface()

Create an interface to MOS The url is constructed from the environment variables MOS_BACKEND_HOST and MOS_BACKEND_HOST. The authentication token is assumed to be available in the environment variable MOS_BACKEND_TOKEN.

MOSInterface.Interface — Method

Interface(url::String)

Create an interface to MOS at the given url. The authentication token is assumed to be available in the environment variable MOS_BACKEND_TOKEN.

MOSInterface.Interface — Method

Interface(url::String, token::String)

Create an interface to MOS at the given url and using the given token for authentication.

MOSInterface.delete_model_with_name — Method

delete_model_with_name(i::Interface, name::String)

Delete model with given name.

MOSInterface.get_model — Method

get_model(i::Interface, id::Integer)::Model

Get model with given id.

MOSInterface.get_model_with_name — Method

get_model_with_name(i::Interface, name::String)::Model

Get model with given name.

MOSInterface.new_model — Method

new_model(i::Interface, filepath::AbstractString)::Model

Create new MOS model from local annotated file.