Model

MOSInterface.Model — Type

MOS model struct.

MOSInterface.delete_results — Method

delete_results(m::Model)

Delete model results.

MOSInterface.get_id — Method

get_id(m::Model)

Get model id.

MOSInterface.get_owner_id — Method

get_owner_id(m::Model)

Get model owner id.

MOSInterface.get_system — Method

get_system(m::Model)

Get model optimization modeling system.

MOSInterface.get_name — Method

get_name(m::Model)

Get model name.

MOSInterface.get_description — Method

get_description(m::Model)

Get model description.

MOSInterface.get_status — Method

get_status(m::Model)

Get model status.

MOSInterface.reload — Method

reload(m::Model)

Reload optimization model by pulling its latest data from MOS.

MOSInterface.run — Function

run(m::Model, blocking::Bool=true, poll_time::Number=1)

Run optimization model.

MOSInterface.set_interface_file — Method

set_interface_file(m::Model, name::AbstractString, filepath::AbstractString)

Set interface file having a given name with data from a local file.

MOSInterface.set_interface_object — Method

set_interface_object(m::Model, name::AbstractString, data::Any)

Set interface object having a given name with JSON-serializable data.

MOSInterface.show_components — Method

show_components(m::Model)

Display model components.

MOSInterface.show_recipe — Method

show_recipe(m::Model)

Show model execution recipe.